push down DefaultTaxiOptimizerParams.reoptimizationTimeStep; make DefaultTaxiOptimizerParams abstract
cause: inheritance of `@StringGetter`s and `@StringSetter`s is not supported by ReflectiveConfigGroup
